rqst : Idle connection cleanup not closing fd.
We are doing just SVC_DESTROY in svc_rqst_clean_func.
It will not close the fd as there will be senital ref.
For client its still valid connection as its not closed, but
its io will hung as we do svc_rqst_xprt_unregister as part of
unlink api now, even if there is senital ref.

Fix is to do SVC_RELEASE after SVC_DESTROY.

Signed-off-by: Gaurav Gangalwar <gaurav.gangalwar@gmail.com>