Change RPCSEC_GSS window size to 256
The current RPCSEC_GSS window size is 32, which is too small and results
in frequent out of sequence errors when client sends multiple requests
in parallel. This can result in auth failures as metioned in this thread:
https://lists.nfs-ganesha.org/archives/list/devel@lists.nfs-ganesha.org/thread/JSADKLKF3GKNCKPFV5NYEJ3ML3YU7UN7/

Changing the window size to 256, along with reducing the value of
tcp_max_slot_table_entries (at client side) mitigates the issue to
a large extent.