Fixes to enabled NFS over RDMA 1>Connection management. 2>Scalable per connection buffer management to use/reuse rdma registered buffers for NFS request/responses/data buffers. 3>Fixes to make rdma_read and rdma_write work. 4>Fixes to make reply_list and write_list work. 5>Limit rdma connections.
We have supporting patch in Ganesha to use rdma registered for read and
readdir responses.
With this patch and supporting Ganesha patch, we should be able to use
NFSv3 and NFSv4.0 over RDMA.
Delegation has to be disabled for NFSv4.0, since callback channel is not
yet supported over RDMA.
Connection disconnect or umount needs some more fixes to do cleanup,
we will fix it as seperate patch.